Add thread_local_stack test for raceconditions.
